argument types for 'argc' method parameters were really inconsistent -- some long, some short.  object_method_typed() calls our functions using longs -- and I saw something strange in the debugger as a result of long-to-short conversion -- so I have changed everything to use longs for argc.  also fixed bug where we were sometimes trashing memory and made it so that we can use jcom_core_loadextern() to instantiate objects with various numbers of and types of arguments.
git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2064 464221e2-0a0f-0410-983b-f8d57d210c17

+++ Externals/common/jcom.sendreceive.h
@@ -10,7 +10,7 @@
 #ifndef __JCOM_SENDRECEIVE_H__
 #define __JCOM_SENDRECEIVE_H__
 
-typedef void (*t_receive_callback)(t_symbol *name, t_symbol *msg, short argc, t_atom *argv);
+typedef void (*t_receive_callback)(t_symbol *name, t_symbol *msg, long argc, t_atom *argv);
 
 /** Receive Object */
 typedef struct _receive{
